more fixes for codestryle more modularization
