+++ bot/TS3Bot.py
@@ -1,72 +1,52 @@
 #!/usr/bin/python
 import binascii  # crc32
 import datetime  # for date strings
-import json
 import logging
 import os  # operating system commands -check if files exist
 import re
 import sqlite3  # Database
 import traceback
 import urllib.parse  # for fetching guild emblems urls
-from threading import RLock
 
 import requests  # to download guild emblems
-import schedule  # Allows auditing of users every X days
 import ts3
 from ts3 import TS3Error
 from ts3.filetransfer import TS3FileTransfer, TS3UploadError
 from ts3.query import TS3QueryError
 
-from bot import TS3Auth
-from bot.Config import Config
-from bot.ts.TS3Facade import TS3Facade
-from bot.ts.ThreadsafeTSConnection import ThreadsafeTSConnection, default_exception_handler, signal_exception_handler
+from bot import Config, TS3Auth
+from bot.command_checker import CommanderChecker
+from bot.db import ThreadSafeDBConnection
+from bot.gw2_api import get_guild_info
+from bot.ts import TS3Facade, ThreadSafeTSConnection, User, default_exception_handler, signal_exception_handler
 from bot.util import StringShortener
 
 LOG = logging.getLogger(__name__)
 
 
-def request(url):
-    response = requests.get(url, headers={"Content-Type": "application/json"})
-
-    if response.status_code == 200:
-        return json.loads(response.content.decode("utf-8"))
-    else:
-        return None
-
-
-#######################################
-
-
-class ThreadsafeDBConnection():
-    def __init__(self, db_name):
-        self.db_name = db_name
-        self.conn = sqlite3.connect(db_name, check_same_thread=False, detect_types=sqlite3.PARSE_DECLTYPES)
-        self.cursor = self.conn.cursor()
-        self.lock = RLock()
-
-
 class Bot:
+
+    def __init__(self, database: ThreadSafeDBConnection, ts_connection: ThreadSafeTSConnection, ts_repository: TS3Facade, config: Config):
+        self._ts_repository = ts_repository
+        self._ts_connection = ts_connection
+        self._config = config
+        self._database_connection = database
+
+        admin_data, _ = self.ts_connection.ts3exec(lambda ts_con: ts_con.query("whoami").first())
+        self.name = admin_data.get('client_login_name')
+        self.client_id = admin_data.get('client_id')
+
+        self.nickname = self.ts_connection.forceRename(config.bot_nickname)
+
+        self.verified_group = config.verified_group
+        self.vgrp_id = self.groupFind(self.verified_group)
+
+        self.c_audit_date = datetime.date.today()
+        self.commander_checker = CommanderChecker(self, config.poll_group_names)
 
     @property
     def ts_connection(self):
         return self._ts_connection
-
-    def __init__(self, db, ts_connection: ThreadsafeTSConnection, ts_repository: TS3Facade, verified_group, config: Config, bot_nickname="TS3BOT"):
-        self._ts_repository: TS3Facade = ts_repository
-        self._ts_connection: ThreadsafeTSConnection = ts_connection
-        self._config: Config = config
-
-        admin_data, _ = self.ts_connection.ts3exec(lambda ts_con: ts_con.query("whoami").first())
-        self.db_name = db
-        self.name = admin_data.get('client_login_name')
-        self.client_id = admin_data.get('client_id')
-        self.nickname = self.ts_connection.forceRename(bot_nickname)
-        self.verified_group = verified_group
-        self.vgrp_id = self.groupFind(verified_group)
-        self.getUserDatabase()
-        self.c_audit_date = datetime.date.today()
-        self.commander_checker = CommanderChecker(self, config.poll_group_names)
 
     # Helps find the group ID for a group name
     def groupFind(self, group_to_find):
@@ -85,8 +65,8 @@ class Bot:
             return False  # User already verified
 
         # Check if user is authenticated in database and if so, re-adds them to the group
-        with self.dbc.lock:
-            current_entries = self.dbc.cursor.execute("SELECT * FROM users WHERE ts_db_id=?", (unique_client_id,)).fetchall()
+        with self._database_connection.lock:
+            current_entries = self._database_connection.cursor.execute("SELECT * FROM users WHERE ts_db_id=?", (unique_client_id,)).fetchall()
             if len(current_entries) > 0:
                 self.setPermissions(unique_client_id)
                 return False
@@ -125,14 +105,14 @@ class Bot:
             LOG.error("Removing permissions failed: %s", err)  # likely due to bad client id
 
     def removePermissionsByGW2Account(self, gw2account):
-        with self.dbc.lock:
-            tsDbIds = self.dbc.cursor.execute("SELECT ts_db_id FROM users WHERE account_name = ?", (gw2account,)).fetchall()
+        with self._database_connection.lock:
+            tsDbIds = self._database_connection.cursor.execute("SELECT ts_db_id FROM users WHERE account_name = ?", (gw2account,)).fetchall()
             for tdi, in tsDbIds:
                 self.removePermissions(tdi)
                 LOG.debug("Removed permissions from %s", tdi)
-            self.dbc.cursor.execute("DELETE FROM users WHERE account_name = ?", (gw2account,))
-            changes = self.dbc.cursor.execute("SELECT changes()").fetchone()[0]
-            self.dbc.conn.commit()
+            self._database_connection.cursor.execute("DELETE FROM users WHERE account_name = ?", (gw2account,))
+            changes = self._database_connection.cursor.execute("SELECT changes()").fetchone()[0]
+            self._database_connection.conn.commit()
             return changes
 
     def getUserDBEntry(self, client_unique_id):
@@ -140,73 +120,40 @@ class Bot:
         Retrieves the DB entry for a unique client ID.
         Is either a dictionary of database-field-names to values, or None if no such entry was found in the DB.
         """
-        with self.dbc.lock:
-            entry = self.dbc.cursor.execute("SELECT * FROM users WHERE ts_db_id=?", (client_unique_id,)).fetchall()
+        with self._database_connection.lock:
+            entry = self._database_connection.cursor.execute("SELECT * FROM users WHERE ts_db_id=?", (client_unique_id,)).fetchall()
             if len(entry) < 1:
                 # user not registered
                 return None
             entry = entry[0]
-            keys = self.dbc.cursor.description
+            keys = self._database_connection.cursor.description
             assert len(entry) == len(keys)
             return dict([(keys[i][0], entry[i]) for i in range(len(entry))])
 
-    def getUserDatabase(self):
-        if os.path.isfile(self.db_name):
-            self.dbc = ThreadsafeDBConnection(self.db_name)  # sqlite3.connect(self.db_name, check_same_thread = False, detect_types = sqlite3.PARSE_DECLTYPES)
-            LOG.info("Loaded User Database...")
-        else:
-            self.dbc = ThreadsafeDBConnection(self.db_name)  # sqlite3.connect(self.db_name, check_same_thread = False, detect_types = sqlite3.PARSE_DECLTYPES)
-            # self.dbc.cursor = self.dbc.conn.cursor()
-            LOG.info("No User Database found...created new database!")
-            with self.dbc.lock:
-                # USERS
-                self.dbc.cursor.execute("CREATE TABLE users(ts_db_id text primary key, account_name text, api_key text, created_date date, last_audit_date date)")
-                # BOT INFO
-                self.dbc.cursor.execute("CREATE TABLE bot_info(version text, last_succesful_audit date)")
-                self.dbc.conn.commit()
-                self.dbc.cursor.execute('INSERT INTO bot_info (version, last_succesful_audit) VALUES (?,?)', (self._config.current_version, datetime.date.today(),))
-                self.dbc.conn.commit()
-                # GUILD INFO
-                self.dbc.cursor.execute('''CREATE TABLE guilds(
-                                guild_id integer primary key autoincrement,
-                                guild_name text UNIQUE,
-                                ts_group text UNIQUE)''')
-                self.dbc.conn.commit()
-
-                # GUILD IGNORES
-                self.dbc.cursor.execute('''CREATE TABLE guild_ignores(
-                                guild_ignore_id integer primary key autoincrement,
-                                guild_id integer,
-                                ts_db_id text,
-                                ts_name text,
-                                FOREIGN KEY(guild_id) REFERENCES guilds(guild_id),
-                                UNIQUE(guild_id, ts_db_id))''')
-                self.dbc.conn.commit()
-
     def TsClientLimitReached(self, gw_acct_name):
-        with self.dbc.lock:
-            current_entries = self.dbc.cursor.execute("SELECT * FROM users WHERE account_name=?", (gw_acct_name,)).fetchall()
+        with self._database_connection.lock:
+            current_entries = self._database_connection.cursor.execute("SELECT * FROM users WHERE account_name=?", (gw_acct_name,)).fetchall()
             return len(current_entries) >= self._config.client_restriction_limit
 
     def addUserToDB(self, client_unique_id, account_name, api_key, created_date, last_audit_date):
-        with self.dbc.lock:
+        with self._database_connection.lock:
             # client_id = self.getActiveTsUserID(client_unique_id)
-            client_exists = self.dbc.cursor.execute("SELECT * FROM users WHERE ts_db_id=?", (client_unique_id,)).fetchall()
+            client_exists = self._database_connection.cursor.execute("SELECT * FROM users WHERE ts_db_id=?", (client_unique_id,)).fetchall()
             if len(client_exists) > 1:
                 LOG.warning("Found multiple database entries for single unique teamspeakid %s.", client_unique_id)
             if len(client_exists) != 0:  # If client TS database id is in BOT's database.
-                self.dbc.cursor.execute("""UPDATE users SET ts_db_id=?, account_name=?, api_key=?, created_date=?, last_audit_date=? WHERE ts_db_id=?""",
-                                        (client_unique_id, account_name, api_key, created_date, last_audit_date, client_unique_id))
+                self._database_connection.cursor.execute("""UPDATE users SET ts_db_id=?, account_name=?, api_key=?, created_date=?, last_audit_date=? WHERE ts_db_id=?""",
+                                                         (client_unique_id, account_name, api_key, created_date, last_audit_date, client_unique_id))
                 LOG.info("Teamspeak ID %s already in Database updating with new Account Name '%s'. (likely permissions changed by a Teamspeak Admin)", client_unique_id, account_name)
             else:
-                self.dbc.cursor.execute("INSERT INTO users ( ts_db_id, account_name, api_key, created_date, last_audit_date) VALUES(?,?,?,?,?)",
-                                        (client_unique_id, account_name, api_key, created_date, last_audit_date))
-            self.dbc.conn.commit()
+                self._database_connection.cursor.execute("INSERT INTO users ( ts_db_id, account_name, api_key, created_date, last_audit_date) VALUES(?,?,?,?,?)",
+                                                         (client_unique_id, account_name, api_key, created_date, last_audit_date))
+            self._database_connection.conn.commit()
 
     def removeUserFromDB(self, client_db_id):
-        with self.dbc.lock:
-            self.dbc.cursor.execute("DELETE FROM users WHERE ts_db_id=?", (client_db_id,))
-            self.dbc.conn.commit()
+        with self._database_connection.lock:
+            self._database_connection.cursor.execute("DELETE FROM users WHERE ts_db_id=?", (client_db_id,))
+            self._database_connection.conn.commit()
 
     # def updateGuildTags(self, client_db_id, auth):
     def updateGuildTags(self, user, auth):
@@ -232,10 +179,11 @@ class Bot:
         # the input for the parameters here are the names of our own server groups on our TS server.
         current_guild_groups = []
         hidden_groups = {}
-        with self.dbc.lock:
-            current_guild_groups = self.dbc.cursor.execute("SELECT ts_group, guild_name FROM guilds WHERE ts_group IN (%s)" % (param,)).fetchall()
+        with self._database_connection.lock:
+            current_guild_groups = self._database_connection.cursor.execute("SELECT ts_group, guild_name FROM guilds WHERE ts_group IN (%s)" % (param,)).fetchall()
             # groups the user doesn't want to wear
-            hidden_groups = set([g[0] for g in self.dbc.cursor.execute("SELECT g.ts_group FROM guild_ignores AS gi JOIN guilds AS g ON gi.guild_id = g.guild_id  WHERE ts_db_id = ?", (uid,))])
+            hidden_groups = set(
+                [g[0] for g in self._database_connection.cursor.execute("SELECT g.ts_group FROM guild_ignores AS gi JOIN guilds AS g ON gi.guild_id = g.guild_id  WHERE ts_db_id = ?", (uid,))])
         # REMOVE STALE GROUPS
         for ggroup, gname in current_guild_groups:
             if ggroup in hidden_groups:
@@ -255,8 +203,8 @@ class Bot:
         # ADD DUE GROUPS
         for g in ingame_member_of:
             ts_group = None
-            with self.dbc.lock:
-                ts_group = self.dbc.cursor.execute("SELECT ts_group FROM guilds WHERE guild_name = ?", (g,)).fetchone()
+            with self._database_connection.lock:
+                ts_group = self._database_connection.cursor.execute("SELECT ts_group FROM guilds WHERE guild_name = ?", (g,)).fetchone()
             if ts_group:
                 ts_group = ts_group[0]  # first and only column, if a row exists
                 if ts_group not in current_group_names:
@@ -281,8 +229,8 @@ class Bot:
     def _auditUsers(self):
         self.c_audit_date = datetime.date.today()  # Update current date everytime run
         self.db_audit_list = []
-        with self.dbc.lock:
-            self.db_audit_list = self.dbc.cursor.execute('SELECT * FROM users').fetchall()
+        with self._database_connection.lock:
+            self.db_audit_list = self._database_connection.cursor.execute('SELECT * FROM users').fetchall()
         for audit_user in self.db_audit_list:
 
             # Convert to single variables
@@ -303,17 +251,17 @@ class Bot:
                     LOG.info("User %s is still on %s. Succesful audit!", audit_account_name, auth.world.get("name"))
                     # self.getTsDatabaseID(audit_ts_id)
                     self.updateGuildTags(User(self.ts_connection, unique_id=audit_ts_id), auth)
-                    with self.dbc.lock:
-                        self.dbc.cursor.execute("UPDATE users SET last_audit_date = ? WHERE ts_db_id= ?", (self.c_audit_date, audit_ts_id,))
-                        self.dbc.conn.commit()
+                    with self._database_connection.lock:
+                        self._database_connection.cursor.execute("UPDATE users SET last_audit_date = ? WHERE ts_db_id= ?", (self.c_audit_date, audit_ts_id,))
+                        self._database_connection.conn.commit()
                 else:
                     LOG.info("User %s is no longer on our server. Removing access....", audit_account_name)
                     self.removePermissions(audit_ts_id)
                     self.removeUserFromDB(audit_ts_id)
 
-        with self.dbc.lock:
-            self.dbc.cursor.execute('INSERT INTO bot_info (last_succesful_audit) VALUES (?)', (self.c_audit_date,))
-            self.dbc.conn.commit()
+        with self._database_connection.lock:
+            self._database_connection.cursor.execute('INSERT INTO bot_info (last_succesful_audit) VALUES (?)', (self.c_audit_date,))
+            self._database_connection.conn.commit()
 
     def broadcastMessage(self):
         broadcast_message = self._config.locale.get("bot_msg_broadcast")
@@ -379,15 +327,6 @@ class Bot:
                 LOG.info("Channel '%s' already exists. This is probably not a problem. Skipping.", newname)
         return 0
 
-    def getGuildInfo(self, guildname):
-        """
-        Lookup guild by name. If such a guild exists (and the API is available)
-        the info as specified on https://wiki.guildwars2.com/wiki/API:2/guild/:id is returned.
-        Else, None is returned.
-        """
-        ids = request("https://api.guildwars2.com/v2/guild/search?name=%s" % (urllib.parse.quote(guildname),))
-        return None if ids is None or len(ids) == 0 else request("https://api.guildwars2.com/v2/guild/%s" % (ids[0]))
-
     def getActiveCommanders(self):
         return self.commander_checker.execute()
 
@@ -408,12 +347,12 @@ class Bot:
         if name is None:
             return INVALID_PARAMETERS
 
-        ginfo = self.getGuildInfo(name)
+        ginfo = get_guild_info(name)
         if ginfo is None:
             return INVALID_GUILD_NAME
 
-        with self.dbc.lock:
-            g = self.dbc.cursor.execute("SELECT ts_group FROM guilds WHERE guild_name = ?", (name,)).fetchone()
+        with self._database_connection.lock:
+            g = self._database_connection.cursor.execute("SELECT ts_group FROM guilds WHERE guild_name = ?", (name,)).fetchone()
             groupname = g[0] if g is not None else None
 
         if groupname is None:
@@ -424,9 +363,9 @@ class Bot:
 
         # FROM DB
         LOG.debug("Deleting guild '%s' from DB.", name)
-        with self.dbc.lock:
-            self.dbc.cursor.execute("DELETE FROM guilds WHERE guild_name = ?", (name,))
-            self.dbc.conn.commit()
+        with self._database_connection.lock:
+            self._database_connection.cursor.execute("DELETE FROM guilds WHERE guild_name = ?", (name,))
+            self._database_connection.conn.commit()
 
         # CHANNEL
         channelname = "%s [%s]" % (name, tag)
@@ -486,7 +425,7 @@ class Bot:
         # lock for the whole block to avoid constant interference
         # locking the ts3conn is vital to properly do the TS3FileTransfer
         # down the line.
-        with ts3conn.lock, self.dbc.lock:
+        with ts3conn.lock, self._database_connection.lock:
             #############################################
             # CHECK IF GROUPS OR CHANNELS ALREADY EXIST #
             #############################################
@@ -498,8 +437,8 @@ class Bot:
                 LOG.debug("Can not create a group '%s', because it already exists. Aborting guild creation.", group)
                 return DUPLICATE_TS_GROUP
 
-            with self.dbc.lock:
-                dbgroups = self.dbc.cursor.execute("SELECT ts_group, guild_name FROM guilds WHERE ts_group = ?", (groupname,)).fetchall()
+            with self._database_connection.lock:
+                dbgroups = self._database_connection.cursor.execute("SELECT ts_group, guild_name FROM guilds WHERE ts_group = ?", (groupname,)).fetchall()
                 if len(dbgroups) > 0:
                     LOG.debug("Can not create a DB entry for TS group '%s', as it already exists. Aborting guild creation.", groupname)
                     return DUPLICATE_DB_ENTRY
@@ -584,9 +523,9 @@ class Bot:
         ###################
         # must exist in DB before creating group to have it available when reordering groups.
         LOG.debug("Creating entry in database for auto assignment of guild group...")
-        with self.dbc.lock:
-            self.dbc.cursor.execute("INSERT INTO guilds(ts_group, guild_name) VALUES(?,?)", (groupname, name))
-            self.dbc.conn.commit()
+        with self._database_connection.lock:
+            self._database_connection.cursor.execute("INSERT INTO guilds(ts_group, guild_name) VALUES(?,?)", (groupname, name))
+            self._database_connection.conn.commit()
 
         #######################
         # CREATE SERVER GROUP #
@@ -623,8 +562,8 @@ class Bot:
 
         groups.append({"sgid": resp.get("sgid"), "name": groupname})  # the newly created group has to be added to properly iterate over the guild groups
 
-        with self.dbc.lock:
-            guildgroups = [g[0] for g in self.dbc.cursor.execute("SELECT ts_group FROM guilds ORDER BY ts_group").fetchall()]
+        with self._database_connection.lock:
+            guildgroups = [g[0] for g in self._database_connection.cursor.execute("SELECT ts_group FROM guilds ORDER BY ts_group").fetchall()]
         for i in range(len(guildgroups)):
             g = next((g for g in groups if g.get("name") == guildgroups[i]), None)
             if g is None:
@@ -649,8 +588,8 @@ class Bot:
             LOG.debug("Can not find a group '%s' for guild contacts. Skipping.", contactgroup)
         else:
             for c in contacts:
-                with self.dbc.lock:
-                    accs = [row[0] for row in self.dbc.cursor.execute("SELECT ts_db_id FROM users WHERE lower(account_name) = lower(?)", (c,)).fetchall()]
+                with self._database_connection.lock:
+                    accs = [row[0] for row in self._database_connection.cursor.execute("SELECT ts_db_id FROM users WHERE lower(account_name) = lower(?)", (c,)).fetchall()]
                     for a in accs:
                         try:
                             u = User(ts3conn, unique_id=a, ex_hand=signal_exception_handler)
@@ -735,7 +674,7 @@ class Bot:
                 f"[hr]\n")
         return text
 
-    def clientMessageHandler(self, ipcserver, clientsocket, message):
+    def clientMessageHandler(self, clientsocket, message):
         mtype = self.try_get(message, "type", lower=True)
         mcommand = self.try_get(message, "command", lower=True)
         margs = self.try_get(message, "args", typer=lambda a: dict(a), default={})
@@ -796,24 +735,25 @@ class Bot:
                 cmd, args = self.commandCheck(raw_cmd)  # sanitize the commands but also restricts commands to a list of known allowed commands
                 if cmd == "hideguild":
                     LOG.info("User '%s' wants to hide guild '%s'.", rec_from_name, args[0])
-                    with self.dbc.lock:
+                    with self._database_connection.lock:
                         try:
-                            self.dbc.cursor.execute("INSERT INTO guild_ignores(guild_id, ts_db_id, ts_name) VALUES((SELECT guild_id FROM guilds WHERE ts_group = ?), ?,?)",
-                                                    (args[0], rec_from_uid, rec_from_name))
-                            self.dbc.conn.commit()
+                            self._database_connection.cursor.execute("INSERT INTO guild_ignores(guild_id, ts_db_id, ts_name) VALUES((SELECT guild_id FROM guilds WHERE ts_group = ?), ?,?)",
+                                                                     (args[0], rec_from_uid, rec_from_name))
+                            self._database_connection.conn.commit()
                             LOG.debug("Success!")
                             self._ts_repository.send_text_message_to_client(rec_from_id, self._config.locale.get("bot_hide_guild_success"))
                         except sqlite3.IntegrityError:
-                            self.dbc.conn.rollback()
+                            self._database_connection.conn.rollback()
                             LOG.debug("Failed. The group probably doesn't exist or the user is already hiding that group.")
                             self._ts_repository.send_text_message_to_client(rec_from_id, self._config.locale.get("bot_hide_guild_unknown"))
 
                 elif cmd == "unhideguild":
                     LOG.info("User '%s' wants to unhide guild '%s'.", rec_from_name, args[0])
-                    with self.dbc.lock:
-                        self.dbc.cursor.execute("DELETE FROM guild_ignores WHERE guild_id = (SELECT guild_id FROM guilds WHERE ts_group = ? AND ts_db_id = ?)", (args[0], rec_from_uid))
-                        changes = self.dbc.cursor.execute("SELECT changes()").fetchone()[0]
-                        self.dbc.conn.commit()
+                    with self._database_connection.lock:
+                        self._database_connection.cursor.execute("DELETE FROM guild_ignores WHERE guild_id = (SELECT guild_id FROM guilds WHERE ts_group = ? AND ts_db_id = ?)",
+                                                                 (args[0], rec_from_uid))
+                        changes = self._database_connection.cursor.execute("SELECT changes()").fetchone()[0]
+                        self._database_connection.conn.commit()
                         if changes > 0:
                             LOG.debug("Success!")
                             self._ts_repository.send_text_message_to_client(rec_from_id, self._config.locale.get("bot_unhide_guild_success"))
@@ -885,161 +825,3 @@ class Bot:
             LOG.error(e)
             LOG.error(traceback.format_exc())
         return None
-
-
-#######################################
-
-class Ticker():
-    """
-    Class that schedules events regularly and wraps the TS3Bot.
-    """
-
-    def __init__(self, ts3bot, interval):
-        self.ts3bot = ts3bot
-        self.interval = interval
-        schedule.every(interval).seconds.do(self.execute)
-
-    def execute(self):
-        pass
-
-
-#######################################
-
-class Channel():
-    def __init__(self, ts_conn, channel_id):
-        self.ts_conn = ts_conn
-        self.channel_id = channel_id
-
-
-#######################################
-
-class User():
-    """
-    Class that interfaces the Teamspeak-API with user-specific calls more convenient.
-    Since calls to the API are penalised, the class also tries to minimise those calls
-    by only resolving properties when they are actually needed and then caching them (if sensible).
-    """
-
-    def __init__(self, ts_conn, unique_id=None, ts_db_id=None, client_id=None, ex_hand=None):
-        self.ts_conn = ts_conn
-        self._unique_id = unique_id
-        self._ts_db_id = ts_db_id
-        self._client_id = client_id
-        self._exception_handler = ex_hand if ex_hand is not None else default_exception_handler
-
-        if all(x is None for x in [unique_id, ts_db_id, client_id]):
-            raise ValueError("At least one ID must be non-null")
-
-    def __repr__(self):
-        return str(self)
-
-    def __str__(self):
-        return "User[unique_id: %s, ts_db_id: %s, client_id: %s]" % (self.unique_id, self.ts_db_id, self._client_id)
-
-    @property
-    def current_channel(self):
-        entry = next((c for c in self.ts_conn.ts3exec(lambda t: t.query("clientlist").all(), self._exception_handler)[0] if c.get("clid") == self.client_id), None)
-        if entry:
-            self._ts_db_id = entry.get("client_database_id")  # since we are already retrieving this information...
-        return Channel(self.ts_conn, entry.get("cid")) if entry else None
-
-    @property
-    def name(self):
-        return self.ts_conn.ts3exec(lambda t: t.query("clientgetids", cluid=self.unique_id).first().get("name"), self._exception_handler)[0]
-
-    @property
-    def unique_id(self):
-        if self._unique_id is None:
-            if self._ts_db_id is not None:
-                self._unique_id, ex = self.ts_conn.ts3exec(lambda t: t.query("clientgetnamefromdbid", cldbid=self._ts_db_id).first().get("cluid"), self._exception_handler)
-            elif self._client_id is not None:
-                ids, ex = self.ts_conn.ts3exec(lambda t: t.query("clientinfo", clid=self._client_id).first(), self._exception_handler)
-                self._unique_id = ids.get("client_unique_identifier")
-                self._ts_db_id = ids.get("client_databased_id")  # not required, but since we already queried it...
-            else:
-                raise ValueError("Unique ID can not be retrieved")
-        return self._unique_id
-
-    @property
-    def ts_db_id(self):
-        if self._ts_db_id is None:
-            if self._unique_id is not None:
-                self._ts_db_id, ex = self.ts_conn.ts3exec(lambda t: t.query("clientgetdbidfromuid", cluid=self._unique_id).first().get("cldbid"), self._exception_handler)
-            elif self._client_id is not None:
-                ids, ex = self.ts_conn.ts3exec(lambda t: t.query("clientinfo", clid=self._client_id).first(), self._exception_handler)
-                self._unique_id = ids.get("client_unique_identifier")  # not required, but since we already queried it...
-                self._ts_db_id = ids.get("client_database_id")
-            else:
-                raise ValueError("TS DB ID can not be retrieved")
-        return self._ts_db_id
-
-    @property
-    def client_id(self):
-        if self._client_id is None:
-            if self._unique_id is not None:
-                # easiest case: unique ID is set
-                self._client_id, ex = self.ts_conn.ts3exec(lambda t: t.query("clientgetids", cluid=self._unique_id).first().get("clid"), self._exception_handler)
-            elif self._ts_db_id is not None:
-                self._unique_id, ex = self.ts_conn.ts3exec(lambda t: t.query("clientgetnamefromdbid", cldbid=self._ts_db_id).first().get("cluid"), self._exception_handler)
-                self._client_id, ex = self.ts_conn.ts3exec(lambda t: t.query("clientgetids", cluid=self._unique_id).first().get("clid"), self._exception_handler)
-            else:
-                raise ValueError("Client ID can not be retrieved")
-        return self._client_id
-
-    #######################################
-
-
-class CommanderChecker():
-    def __init__(self, ts3bot, commander_group_names):
-        self.commander_group_names = commander_group_names
-        self.ts3bot = ts3bot
-
-        cgroups = list(filter(lambda g: g.get("name") in commander_group_names, self.ts3bot.ts_connection.ts3exec(lambda t: t.query("channelgrouplist").all())[0]))
-        if len(cgroups) < 1:
-            LOG.info("Could not find any group of %s to determine commanders by. Disabling this feature.", str(commander_group_names))
-            self.commander_groups = []
-            return
-
-        self.commander_groups = [c.get("cgid") for c in cgroups]
-
-    def execute(self):
-        if not self.commander_groups:
-            return  # disabled if no groups were found
-
-        active_commanders = []
-
-        def retrieve_commanders(tsc):
-            command = tsc.query("channelgroupclientlist")
-            for cgid in self.commander_groups:
-                command.pipe(cgid=cgid)
-            return command.all()
-
-        acs, ts3qe = self.ts3bot.ts_connection.ts3exec(retrieve_commanders, signal_exception_handler)
-        if ts3qe:  # check for .resp, could be another exception type
-            if ts3qe.resp is not None:
-                if ts3qe.resp.error["id"] != "1281":
-                    # 1281 is "database empty result set", which is an expected error
-                    # if not a single user currently wears a tag.
-                    LOG.error("Error while trying to resolve active commanders: %s.", str(ts3qe))
-            else:
-                LOG.error(ts3qe)
-        else:
-            active_commanders_entries = [(c, self.ts3bot.getUserDBEntry(self.ts3bot.getTsUniqueID(c.get("cldbid")))) for c in acs]
-            for ts_entry, db_entry in active_commanders_entries:
-                if db_entry is not None:  # or else the user with the commander group was not registered and therefore not in the DB
-                    u = User(self.ts3bot.ts_connection, ts_db_id=ts_entry.get("cldbid"))
-                    if u.current_channel.channel_id == ts_entry.get("cid"):
-                        # user could have the group in a channel but not be in there atm
-                        ac = {}
-                        ac["account_name"] = db_entry["account_name"]
-                        ac["ts_cluid"] = db_entry["ts_db_id"]
-                        ac["ts_display_name"], ex1 = self.ts3bot.ts_connection.ts3exec(
-                            lambda t, cluid=db_entry["ts_db_id"]: t.query("clientgetnamefromuid", cluid).first().get("name"))  # no, there is probably no easier way to do this. I checked.
-                        ac["ts_channel_name"], ex2 = self.ts3bot.ts_connection.ts3exec(lambda t, cid=ts_entry.get("cid"): t.query("channelinfo", cid=cid).first().get("channel_name"))
-                        if ex1 or ex2:
-                            LOG.warning("Could not determine information for commanding user with ID %s: '%s'. Skipping.", str(ts_entry), ", ".join([str(e) for e in [ex1, ex2] if e is not None]))
-                        else:
-                            active_commanders.append(ac)
-        return {"commanders": active_commanders}
-
-#######################################
