+++ bot/__init__.py
@@ -1,4 +1,4 @@
-from .Config import Config
 from .TS3Bot import Bot
+from .config import Config
 
 __all__ = ['Config', 'Bot']
