+++ bot/db/__init__.py
@@ -1,0 +1,3 @@
+from .database import ThreadSafeDBConnection, get_or_create_database
+
+__all__ = ["ThreadSafeDBConnection", "get_or_create_database"]
