+++ bot/rest/controller/__init__.py
@@ -1,5 +1,6 @@
 from .guild_controller import GuildController
 from .health_controller import HealthController
 from .other_controller import OtherController
+from .reset_roster_controller import ResetRosterController
 
-__all__ = ['GuildController', 'HealthController', 'OtherController']
+__all__ = ['GuildController', 'HealthController', 'OtherController', 'ResetRosterController']
