+++ bot/ts/__init__.py
@@ -1,8 +1,11 @@
 from .TS3Facade import Channel, TS3Facade
-from .ThreadsafeTSConnection import ThreadsafeTSConnection, ignore_exception_handler, signal_exception_handler
+from .ThreadSafeTSConnection import ThreadSafeTSConnection, default_exception_handler, ignore_exception_handler, \
+    signal_exception_handler
+from .user import User
 
 __all__ = [
     'Channel', 'TS3Facade',
-    'ThreadsafeTSConnection',
-    'ignore_exception_handler', 'signal_exception_handler'
+    'ThreadSafeTSConnection',
+    'ignore_exception_handler', 'signal_exception_handler', 'default_exception_handler',
+    'User'
 ]
