+++ bot/util/StringShortener.py
@@ -4,57 +4,57 @@ from functools import cmp_to_key
 from typing import List
 
 
-class StringShortener(object):
+class StringShortener:
     @staticmethod
-    def remove_tags(s, length):
+    def remove_tags(string, length):
         """
         Removes all guild tags, that is all groups of letters in brackets
         at the start of the string. "[x][y] [z] Name" -> "Name"
         """
-        return re.search(r"^(?:\[.*\])*(.*)$", s).group(1).strip()
+        return re.search(r"^(?:\[.*\])*(.*)$", string).group(1).strip()
 
     @staticmethod
-    def only_some_words(s, length):
+    def only_some_words(string, length):
         """
         Splits the string at whitespaces and removes as little
         tokens from the end as needed to match the required string length.
         """
-        ts = s.split()
+        ts = string.split()
         i = 0
         res = ""
         while i < len(ts) and len(res + ts[i]) + i <= length:
             res = "%s %s" % (res, ts[i])
             i += 1
         res = res.strip()
-        return res if len(res) > 0 else s
+        return res if len(res) > 0 else string
 
     @staticmethod
-    def only_first_word(s, length):
+    def only_first_word(string, length):
         """
         Only uses the substring up to the first whitespace.
         """
-        res = re.search(r"^(\S+)", s)
-        return res.group(1).strip() if res is not None else s
+        res = re.search(r"^(\S+)", string)
+        return res.group(1).strip() if res is not None else string
 
     @staticmethod
-    def ellipsis(s, length):
+    def ellipsis(string, length):
         """
         Cuts off the string (even in the middle of a word)
         and puts in an ellipsis at the end.
         """
         assert length > 3
         dots = "..."
-        return "%s%s" % (s[:(length - len(dots))], dots) if len(s) > length else s
+        return "%s%s" % (string[:(length - len(dots))], dots) if len(string) > length else string
 
     @staticmethod
-    def force(s, length):
+    def force(string, length):
         """
         Hard cutoff to a substring of a given length. Ultima ratio.
         """
-        return s[:length] if len(s) > length else s
+        return string[:length] if len(string) > length else string
 
     @staticmethod
-    def _shorten(strings, fs, length, spacer=", "):
+    def _shorten(strings, functions, length, spacer=", "):
         """
         Shortens a list of strings in various, increasingly intrusive ways.
         Tries to only modify the strings as much as needed and as little as possible.
@@ -67,22 +67,24 @@ class StringShortener(object):
         """
         if len(strings) == 0:
             return ""  # early bail
-        sl = math.floor((length - (len(strings) - 1) * len(spacer)) / len(strings))  # devide length evenly, substract some space for the glue-string
-        ss = sorted(zip(strings, range(0, len(strings))),  # zip strings with their original position in the input [(x_1, 0), (x_2, 1), ... (x_n, n-1)]
-                    key=cmp_to_key(lambda x, y: len(x[0]) - len(y[0])))  # order ascending by string length
+        desired_len_each = math.floor((length - (len(strings) - 1) * len(spacer)) / len(strings))  # devide length evenly, substract some space for the glue-string
+        zipped_strings = sorted(zip(strings, range(0, len(strings))),  # zip strings with their original position in the input [(x_1, 0), (x_2, 1), ... (x_n, n-1)]
+                                key=cmp_to_key(lambda x, y: len(x[0]) - len(y[0])))  # order ascending by string length
 
-        # l = sl
-        for i in range(len(ss)):
-            j = 0
-            s, p = ss[i]
-            while len(s) > sl:
-                j_ = fs[j]
-                s = j_(s, sl)
-                ss[i] = (s, p)
-                j += 1  # j is not checked to be within bounds. This means the last function MUST succeed in shortening the string sufficiently!
-            # l = sl + sl - len(s)  # list is ordered by string length, so the earlier names could already be below the limit. In that case, they can "pass on" they characters they did not need.
+        # l = desired_len_each
+        for i in range(len(zipped_strings)):
+            next_function_i = 0
+            string, position = zipped_strings[i]
+            while len(string) > desired_len_each:
+                string = functions[next_function_i](string, desired_len_each)
+                zipped_strings[i] = (string, position)
+                next_function_i += 1  # next_function_i is not checked to be within bounds.
+                # This means the last function MUST succeed in shortening the string sufficiently!
 
-        res = [s for s, p in sorted(ss, key=cmp_to_key(lambda x, y: x[1] - y[1]))]
+            # l = desired_len_each + desired_len_each - len(string)  # list is ordered by string length,
+            # so the earlier names could already be below the limit. In that case, they can "pass on" they characters they did not need.
+
+        res = [s for s, p in sorted(zipped_strings, key=cmp_to_key(lambda x, y: x[1] - y[1]))]
         assert len(spacer.join(res)) <= length
         return res
